Refactor tests & middleware; add coverage for geo/rate limit
- Added and reorganized unit tests for GeographicRestrictionMiddleware and RateLimitingMiddleware, improving coverage and clarity.
- Refactored test code for readability, using modern C# features and stronger assertions.
- Moved ModuleInfo to its own file and namespace.
- Updated technical debt doc with missing Refit interface item.
- Improved middleware and options for resilience, CORS, and edge auth.
- Updated YARP cluster config and enhanced code comments and style.
