+++ src/Bootstrapper/MeAjudaAi.Gateway/Middlewares/ResilientForwarderHttpClientFactory.cs
@@ -1,24 +1,19 @@
-using System.Net;
+using MeAjudaAi.Gateway.Handlers;
 using MeAjudaAi.Gateway.Options;
 using Microsoft.Extensions.Options;
 using Yarp.ReverseProxy.Forwarder;
 
 namespace MeAjudaAi.Gateway.Middlewares;
 
-public class ResilientForwarderHttpClientFactory : IForwarderHttpClientFactory
+/// <summary>
+/// Fábrica customizada para criar HttpMessageInvoker/HttpMessageHandler com políticas de
+/// resiliência (retry, timeout) aplicadas ao encaminhamento de requisições pelo YARP.
+/// </summary>
+public class ResilientForwarderHttpClientFactory(
+    IOptions<GatewayResilienceOptions> options,
+    ILogger<ResilientForwarderHttpClientFactory> logger) : IForwarderHttpClientFactory
 {
-    private readonly GatewayResilienceOptions _options;
-    private readonly ILogger<ResilientForwarderHttpClientFactory> _logger;
-    private static readonly HashSet<string> DefaultRetryableMethods =
-        new(["GET", "HEAD", "OPTIONS"], StringComparer.OrdinalIgnoreCase);
-
-    public ResilientForwarderHttpClientFactory(
-        IOptions<GatewayResilienceOptions> options,
-        ILogger<ResilientForwarderHttpClientFactory> logger)
-    {
-        _options = options.Value;
-        _logger = logger;
-    }
+    private readonly GatewayResilienceOptions _options = options.Value;
 
     public HttpMessageInvoker CreateClient(ForwarderHttpClientContext context)
     {
@@ -29,7 +24,7 @@ public class ResilientForwarderHttpClientFactory : IForwarderHttpClientFactory
 
     public HttpMessageHandler CreateHandler(ForwarderHttpClientContext context)
     {
-        _logger.LogDebug(
+        logger.LogDebug(
             "Creating resilient HttpHandler with timeout {Timeout}s and retry count {RetryCount}",
             _options.TimeoutSeconds,
             _options.RetryCount);
@@ -46,106 +41,9 @@ public class ResilientForwarderHttpClientFactory : IForwarderHttpClientFactory
 
         if (_options.RetryCount > 0)
         {
-            return new RetryDelegatingHandler(_options, _logger) { InnerHandler = socketsHandler };
+            return new RetryDelegatingHandler(_options, logger) { InnerHandler = socketsHandler };
         }
 
         return socketsHandler;
     }
-}
-
-internal sealed class RetryDelegatingHandler : DelegatingHandler
-{
-    private readonly GatewayResilienceOptions _options;
-    private readonly ILogger<ResilientForwarderHttpClientFactory> _logger;
-    private static readonly HashSet<string> DefaultRetryableMethods =
-        new(["GET", "HEAD", "OPTIONS"], StringComparer.OrdinalIgnoreCase);
-
-    public RetryDelegatingHandler(GatewayResilienceOptions options, ILogger<ResilientForwarderHttpClientFactory> logger)
-    {
-        _options = options;
-        _logger = logger;
-    }
-
-    protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken ct)
-    {
-        using var timeoutCts = CancellationTokenSource.CreateLinkedTokenSource(ct);
-        timeoutCts.CancelAfter(TimeSpan.FromSeconds(_options.TimeoutSeconds));
-
-        var retryableMethods = _options.RetryableMethods?.Count > 0 
-            ? _options.RetryableMethods 
-            : DefaultRetryableMethods.ToList();
-        
-        var allowRetry = retryableMethods.Contains(request.Method.Method, StringComparer.OrdinalIgnoreCase);
-        
-        if (!allowRetry || _options.RetryCount <= 0)
-        {
-            return await base.SendAsync(request, timeoutCts.Token);
-        }
-
-        HttpResponseMessage? last = null;
-        Exception? lastException = null;
-        
-        for (int attempt = 0; attempt <= _options.RetryCount; attempt++)
-        {
-            try
-            {
-                last = await base.SendAsync(request, timeoutCts.Token);
-                
-                if (!IsTransient(last))
-                {
-                    return last;
-                }
-
-                _logger.LogWarning(
-                    "Retry attempt {AttemptNumber}/{MaxAttempts} for {Method} {Url} - Status: {StatusCode}",
-                    attempt + 1,
-                    _options.RetryCount,
-                    request.Method.Method,
-                    request.RequestUri,
-                    last.StatusCode);
-
-                if (attempt < _options.RetryCount)
-                {
-                    last.Dispose();
-                    last = null;
-                }
-            }
-            catch (Exception ex) when (IsTransientException(ex))
-            {
-                lastException = ex;
-                _logger.LogWarning(
-                    "Retry attempt {AttemptNumber}/{MaxAttempts} for {Method} {Url} - Exception: {Message}",
-                    attempt + 1,
-                    _options.RetryCount,
-                    request.Method.Method,
-                    request.RequestUri,
-                    ex.Message);
-            }
-
-            if (attempt < _options.RetryCount)
-            {
-                var delay = TimeSpan.FromMilliseconds(_options.RetryBaseDelayMs * Math.Pow(2, attempt));
-                await Task.Delay(delay, timeoutCts.Token);
-            }
-        }
-
-        if (last != null)
-        {
-            return last;
-        }
-
-        if (lastException != null)
-        {
-            throw lastException;
-        }
-
-        throw new HttpRequestException("All retry attempts failed");
-    }
-
-    private static bool IsTransient(HttpResponseMessage response) =>
-        (int)response.StatusCode >= 500 ||
-        response.StatusCode is HttpStatusCode.TooManyRequests or HttpStatusCode.RequestTimeout or HttpStatusCode.GatewayTimeout;
-
-    private static bool IsTransientException(Exception ex) =>
-        ex is HttpRequestException or TaskCanceledException or IOException;
 }
