+++ src/Bootstrapper/MeAjudaAi.Gateway/Options/EdgeAuthGuardOptions.cs
@@ -1,11 +1,15 @@
 namespace MeAjudaAi.Gateway.Options;
 
+/// <summary>
+/// Opções de configuração para o middleware de proteção de autenticação de borda (EdgeAuthGuard).
+/// </summary>
 public class EdgeAuthGuardOptions
 {
     public const string SectionName = "EdgeAuthGuard";
 
     public bool Enabled { get; set; } = true;
-    public List<string> PublicRoutes { get; set; } = [
+    public List<string> PublicRoutes { get; set; } = new List<string>
+    {
         "/health",
         "/swagger",
         "/api/v1/auth/login",
@@ -14,7 +18,7 @@ public class EdgeAuthGuardOptions
         "/api/v1/customers/register",
         "/api/v1/providers/register",
         "/webhooks/stripe"
-    ];
+    };
     public string ChallengeHeader { get; set; } = "X-Gateway-Challenge";
     public string AuthenticatedHeader { get; set; } = "X-Gateway-Authenticated";
 }
