+++ src/Bootstrapper/MeAjudaAi.Gateway/Options/GatewayResilienceOptions.cs
@@ -1,5 +1,8 @@
 namespace MeAjudaAi.Gateway.Options;
 
+/// <summary>
+/// Opções de configuração para as políticas de resiliência aplicadas no Gateway.
+/// </summary>
 public class GatewayResilienceOptions
 {
     public const string SectionName = "GatewayResilience";
@@ -7,5 +10,5 @@ public class GatewayResilienceOptions
     public int TimeoutSeconds { get; set; } = 30;
     public int RetryCount { get; set; } = 3;
     public int RetryBaseDelayMs { get; set; } = 100;
-    public List<string> RetryableMethods { get; set; } = ["GET", "HEAD", "OPTIONS"];
-}
+    public List<string> RetryableMethods { get; set; } = new List<string> { "GET", "HEAD", "OPTIONS" };
+    }
