+++ src/Bootstrapper/MeAjudaAi.Gateway/appsettings.json
@@ -62,15 +62,17 @@
     },
     "Clusters": {
       "api-cluster": {
-        "HttpClient": {
+        "HttpRequest": {
           "ActivityTimeout": "00:00:45"
         },
         "HealthCheck": {
-          "Enabled": true,
-          "Interval": "00:00:10",
-          "Timeout": "00:00:05",
-          "Policy": "ConsecutiveFailures",
-          "Path": "/health"
+          "Active": {
+            "Enabled": true,
+            "Interval": "00:00:10",
+            "Timeout": "00:00:05",
+            "Policy": "ConsecutiveFailures",
+            "Path": "/health"
+          }
         },
         "Destinations": {
           "destination1": {
