+++ tests/MeAjudaAi.ApiService.Tests/Infrastructure/Authentication/TestAuthenticationHandler.cs
@@ -1,8 +1,8 @@
-using System.Security.Claims;
-using System.Text.Encodings.Web;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using System.Security.Claims;
+using System.Text.Encodings.Web;
 
 namespace MeAjudaAi.ApiService.Tests.Infrastructure.Authentication;
