+++ tests/MeAjudaAi.ApiService.Tests/Unit/Middlewares/RequestLoggingMiddlewareTests.cs
@@ -1,9 +1,9 @@
-using System.Security.Claims;
 using FluentAssertions;
 using MeAjudaAi.ApiService.Middlewares;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Moq;
+using System.Security.Claims;
 using Xunit;
 
 namespace MeAjudaAi.ApiService.Tests.Unit.Middlewares;
