+++ tests/MeAjudaAi.Architecture.Tests/Helpers/ArchitecturalDiscoveryHelper.cs
@@ -1,5 +1,5 @@
-using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
+using System.Reflection;
 
 namespace MeAjudaAi.Architecture.Tests.Helpers;
