+++ tests/MeAjudaAi.Architecture.Tests/NamingConventionTests.cs
@@ -1,5 +1,6 @@
-using System.Reflection;
 using MeAjudaAi.Architecture.Tests.Helpers;
+using MeAjudaAi.Architecture.Tests.Helpers.Models;
+using System.Reflection;
 
 namespace MeAjudaAi.Architecture.Tests;
