+++ tests/MeAjudaAi.Gateway.Tests/Unit/Middleware/GeographicRestrictionErrorResponseTests.cs
@@ -1,0 +1,54 @@
+using FluentAssertions;
+using MeAjudaAi.Shared.Middleware;
+
+namespace MeAjudaAi.Gateway.Tests.Unit.Middleware;
+
+[Trait("Category", "Unit")]
+[Trait("Layer", "Gateway")]
+public class GeographicRestrictionErrorResponseTests
+{
+    [Fact]
+    public void GeographicRestrictionErrorResponse_ShouldCreateCorrectly()
+    {
+        var response = new GeographicRestrictionErrorResponse(
+            "Access denied",
+            UserLocation.Create("São Paulo", "SP"),
+            [AllowedCity.Create("São Paulo", "SP")],
+            ["SP", "RJ"]
+        );
+
+        response.message.Should().Be("Access denied");
+        response.userLocation.Should().NotBeNull();
+        response.userLocation.City.Should().Be("São Paulo");
+        response.userLocation.State.Should().Be("SP");
+        response.allowedCities.Should().HaveCount(1);
+        response.allowedStates.Should().HaveCount(2);
+    }
+
+    [Fact]
+    public void AllowedCity_Create_ShouldSetPropertiesCorrectly()
+    {
+        var allowedCity = AllowedCity.Create("São Paulo", "SP");
+
+        allowedCity.Name.Should().Be("São Paulo");
+        allowedCity.State.Should().Be("SP");
+    }
+
+    [Fact]
+    public void UserLocation_Create_ShouldSetPropertiesCorrectly()
+    {
+        var userLocation = UserLocation.Create("São Paulo", "SP");
+
+        userLocation.City.Should().Be("São Paulo");
+        userLocation.State.Should().Be("SP");
+    }
+
+    [Fact]
+    public void UserLocation_Create_WithNullValues_ShouldWork()
+    {
+        var userLocation = UserLocation.Create(null, null);
+
+        userLocation.City.Should().BeNull();
+        userLocation.State.Should().BeNull();
+    }
+}
