+++ tests/MeAjudaAi.Gateway.Tests/Unit/Middleware/ResilientForwarderHttpClientFactoryTests.cs
@@ -1,10 +1,10 @@
-using System.Net;
 using FluentAssertions;
 using MeAjudaAi.Gateway.Middlewares;
 using MeAjudaAi.Gateway.Options;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Moq;
+using System.Net;
 using Yarp.ReverseProxy.Forwarder;
 
 namespace MeAjudaAi.Gateway.Tests.Unit.Middleware;
