+++ tests/MeAjudaAi.Integration.Tests/Modules/Locations/LocationsApiIntegrationTests.cs
@@ -2,7 +2,7 @@ using System.Net;
 using System.Net.Http.Json;
 using FluentAssertions;
 using MeAjudaAi.Integration.Tests.Base;
-using MeAjudaAi.Contracts.Contracts.Modules.Locations.DTOs;
+using MeAjudaAi.Contracts.Modules.Locations.DTOs;
 
 namespace MeAjudaAi.Integration.Tests.Modules.Locations;
